routers moved to routers.py. connected routers.py to main.py
